requirements.txt
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.6.15
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-WTF==1.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
python-dotenv==1.1.1
sniffio==1.3.1
typing_extensions==4.14.1
Werkzeug==3.1.3
WTForms==3.2.1
